requirements.txt
streamlit
openai
streamlit-mic-recorder
opencv-python-headless
numpy
requests
Pillow
watchdog